requirements.txt
Pillow==8.2.0
streamlit==0.80.0